refactor(dh): switch to using wizard version for create user role (#1613)
* feat(dh): remove create-userrole route

* feat(dh): generate create user role modal library

* chore(dh): add missing property to project files

* chore(dh): adjust config files

* chore(dh): change method parameters to be more specific

* refactor(dh): switch to exhaustMap operator

* feat(dh): create user role as a wizard in a modal

* feat(dh): switch to using wizard version for create user role

* feat(dh): remove legacy create user role library

* chore(dh): adjust translations

* chore: add license

* Chore: bump version to 0.3.214 for build/infrastructure/eo/chart/Chart.yaml

* chore(dh): remove old translations

* Chore: bump version to 0.3.214 for build/infrastructure/eo/chart/values.yaml

* chore(dh): delete unused method

* chore(dh): rename library

---------
