+++ libs/dh/admin/data-access-graphql/project.json
@@ -1,5 +1,6 @@
 {
   "name": "dh-admin-data-access-graphql",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "sourceRoot": "libs/dh/admin/data-access-graphql/src",
   "prefix": "dh",
