+++ libs/dh/admin/feature-create-user-role/project.json
@@ -1,6 +1,6 @@
 {
   "name": "dh-admin-feature-create-user-role",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "$schema": "../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "sourceRoot": "libs/dh/admin/feature-create-user-role/src",
   "prefix": "dh",
@@ -15,6 +15,7 @@
     },
     "lint": {
       "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
       "options": {
         "lintFilePatterns": [
           "libs/dh/admin/feature-create-user-role/**/*.ts",
