--- libs/dh/admin/feature-create-user-role/src/lib/tabs/create-role-permissions-tab/dh-create-userrole-permissions-tab.component.html
@@ -1,39 +1,0 @@
-<!--
-@license
-Copyright 2020 Energinet DataHub A/S
-
-Licensed under the Apache License, Version 2.0 (the "License2");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<ng-container *transloco="let t; read: 'admin.userManagement.createrole.tabs.permissions'">
-  <watt-card>
-    <watt-card-title>
-      <h4>
-        {{ t("title") }}
-      </h4>
-    </watt-card-title>
-
-    <dh-permissions-table
-      [permissions]="permissions"
-      (selectionChanged)="onSelectionChange($event)"
-    ></dh-permissions-table>
-
-    <!-- Empty -->
-    <div *ngIf="permissions.length === 0" class="no-content-container">
-      <watt-empty-state
-        icon="custom-power"
-        size="small"
-        [title]="t('noPermissions')"
-      ></watt-empty-state>
-    </div>
-  </watt-card>
-</ng-container>
