+++ libs/dh/admin/routing/src/lib/dh-admin-paths.ts
@@ -16,4 +16,3 @@
  */
 export const dhAdminPath = 'admin';
 export const dhAdminUserManagementPath = 'user-management';
-export const dhAdminUserRoleManagementCreatePath = 'create-userrole';
