+++ libs/dh/admin/shell/src/lib/dh-admin-shell.routes.ts
@@ -17,10 +17,7 @@
 import { Routes } from '@angular/router';
 
 import { PermissionGuard } from '@energinet-datahub/dh/shared/feature-authorization';
-import {
-  dhAdminUserManagementPath,
-  dhAdminUserRoleManagementCreatePath,
-} from '@energinet-datahub/dh/admin/routing';
+import { dhAdminUserManagementPath } from '@energinet-datahub/dh/admin/routing';
 
 export const routes: Routes = [
   {
@@ -42,14 +39,6 @@ export const routes: Routes = [
         data: {
           titleTranslationKey: 'admin.userManagement.topBarTitle',
         },
-      },
-      {
-        path: dhAdminUserRoleManagementCreatePath,
-        canActivate: [PermissionGuard(['users:manage'])],
-        loadComponent: () =>
-          import('@energinet-datahub/dh/admin/feature-create-user-role').then(
-            (m) => m.DhAdminCreateUserRoleComponent
-          ),
       },
     ],
   },
