+++ libs/dh/globalization/assets-localization/src/assets/i18n/da.json
@@ -1041,24 +1041,22 @@
       "createrole": {
         "headline": "Ny brugerrolle",
         "save": "Opret brugerrolle",
-        "cancel": "Annuller",
         "createRoleRequest": {
           "start": "Opretter brugerrolle",
           "error": "Der opstod desværre en fejl, da vi skulle oprette brugerrollen.",
           "success": "Brugerrollen er oprettet"
         },
-        "tabs": {
-          "masterdata": {
-            "tabLabel": "Stamdata",
-            "nameMaxLengthErrorText": "Navn skal være under 250 karakterer",
-            "name": "Navn",
-            "description": "Beskrivelse",
+        "steps": {
+          "masterData": {
+            "label": "Stamdata",
             "eicFunction": "Markedsrolle",
-            "selectEicFunctionStatus": "Vælg markedsrolle"
+            "selectEicFunctionStatus": "Vælg markedsrolle",
+            "name": "Navn",
+            "nameMaxLengthErrorText": "Navn skal være under 250 karakterer",
+            "description": "Beskrivelse"
           },
           "permissions": {
-            "tabLabel": "Rettigheder",
-            "title": "Rettigheder",
+            "label": "Tildel rettigheder",
             "noPermissions": "Ingen rettigheder fundet til den valgte markedsrolle"
           }
         }
