+++ libs/dh/shared/util-application-insights/project.json
@@ -1,5 +1,6 @@
 {
   "name": "dh-shared-util-application-insights",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "sourceRoot": "libs/dh/shared/util-application-insights/src",
   "prefix": "dh",
