+++ libs/dh/wholesale/data-access-api/project.json
@@ -1,5 +1,6 @@
 {
   "name": "dh-wholesale-data-access-api",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "sourceRoot": "libs/dh/wholesale/data-access-api/src",
   "prefix": "dh",
