+++ libs/eo/certificates/project.json
@@ -1,5 +1,6 @@
 {
   "name": "eo-certificates",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "sourceRoot": "libs/eo/certificates/src",
   "prefix": "eo",
